agentic-health-monitor/backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
sqlalchemy
openai
chromadb
supabase
